remove any manual async code
I need to keep reminding myself...

Threads *are* continuations.

If OS-level threads are truly too expensive,
we can implement green threads if we absolutely must.

It doesn't even have to be visible to the user!
We can make a wrapper that uses green threads and abstract it behind a
user-friendly "async" interface if they insist so.