impl insert
It's broken, but I just realized that using nibbles is totally redundant
since under realistic usage we'll *never* see common prefixes.

The only situation where this can occur is a form of HashDOS.
For that the proper solution is an additional hashing step,
though as long as we use a tree HashDOS ought not to be an issue.
It will only cause mild performance degradation due to increased point
chasing, which is O(log n) at worst (and not O(n)).